+++ src/controllers/cohort.js
@@ -1,4 +1,5 @@
 import { createCohort, getCohorts, getCohort } from '../domain/cohort.js'
+import { CohortExercise } from '../domain/cohortExercise.js'
 import { sendDataResponse, sendMessageResponse } from '../utils/responses.js'
 import User from '../domain/user.js'
 
@@ -49,5 +50,25 @@ export const getCohortById = async (req, res) => {
     return sendDataResponse(res, 201, cohort)
   } catch (e) {
     return sendMessageResponse(res, 500, 'Unable to get cohort')
+  }
+}
+
+export const getAllCohortExercises = async (req, res) => {
+  const cohortId = Number(req.params.id)
+  const whereData = {}
+  whereData.cohortId = cohortId
+
+  try {
+    const foundExercise = await CohortExercise.findAll({ whereData })
+
+    const formattedExercise = foundExercise.map((exercise) => {
+      return {
+        ...exercise.toJSON().exercise
+      }
+    })
+    // console.log('formattedExercise : ',formattedExercise)
+    return sendDataResponse(res, 200, { notes: formattedExercise })
+  } catch (e) {
+    return sendMessageResponse(res, 500, 'Unable to get exercise')
   }
 }
