+++ src/routes/cohort.js
@@ -1,5 +1,10 @@
 import { Router } from 'express'
-import { create, cohorts, getCohortById } from '../controllers/cohort.js'
+import {
+  create,
+  cohorts,
+  getCohortById,
+  getAllCohortExercises
+} from '../controllers/cohort.js'
 import {
   validateAuthentication,
   validateTeacherRole
@@ -10,5 +15,11 @@ const router = Router()
 router.post('/', validateAuthentication, validateTeacherRole, create)
 router.get('/', validateAuthentication, validateTeacherRole, cohorts)
 router.get('/:id', validateAuthentication, validateTeacherRole, getCohortById)
+
+router.get(
+  '/:id/cohortExercises',
+  validateAuthentication,
+  getAllCohortExercises
+)
 
 export default router
